[pcr] Join files pkg/pcr/get_measurements{,_pcr0}.go
